+++ gloo/benchmark/main.cc
@@ -19,6 +19,7 @@
 #include "gloo/broadcast_one_to_all.h"
 #include "gloo/pairwise_exchange.h"
 #include "gloo/reduce_scatter.h"
+#include "gloo/common/aligned_allocator.h"
 #include "gloo/common/common.h"
 #include "gloo/common/logging.h"
 #include "gloo/context.h"
